Add function for reset button
