+++ js/user-form.js
@@ -1,3 +1,6 @@
+import {validateTimeIn,  validatePrices,  validateRooms} from './valid-form.js';
+import {mainPinMarker, map} from './map.js';
+import {LAT_DEFAULT, LNG_DEFAULT} from './data.js';
 import {sendData} from './api.js';
 const adForm = document.querySelector('.ad-form');
 
@@ -13,4 +16,15 @@ const setUserFormSubmit = (onSuccess, onError) => {
   });
 };
 
-export {setUserFormSubmit};
+const resetForm = (lat, lng) => {
+  document.querySelector('.map__filters').reset();
+  document.querySelector('.ad-form').reset();
+  document.querySelector('#address').value = `${lat.toFixed(5)}, ${lng.toFixed(5)}`;
+  mainPinMarker.setLatLng({lat: LAT_DEFAULT, lng: LNG_DEFAULT});
+  map.setView({lat: LAT_DEFAULT, lng: LNG_DEFAULT}, 12);
+  validateTimeIn();
+  validatePrices();
+  validateRooms();
+};
+
+export {setUserFormSubmit, resetForm};
